+++ src/main/java/lotto/domain/CalculationResult.java
@@ -1,4 +1,4 @@
-package lotto.mainframe;
+package lotto.domain;
 
 import java.util.List;
 import lotto.view.GameOutput;
