+++ src/main/java/lotto/domain/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.mainframe;
+package lotto.domain;
 
 import java.util.HashSet;
 import java.util.List;
