+++ src/main/java/lotto/domain/LottoNumberGeneration.java
@@ -1,4 +1,4 @@
-package lotto.mainframe;
+package lotto.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
