+++ src/main/java/lotto/domain/User.java
@@ -1,4 +1,4 @@
-package lotto.mainframe;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.Arrays;
